+++ Classes/ImageService/ImgixImageService.php
@@ -1,0 +1,109 @@
+<?php
+
+declare(strict_types=1);
+
+namespace SomehowDigital\Typo3\MediaProcessing\ImageService;
+
+use SomehowDigital\Typo3\MediaProcessing\UriBuilder\ImgixUri;
+use SomehowDigital\Typo3\MediaProcessing\UriBuilder\UriSourceInterface;
+use TYPO3\CMS\Core\Imaging\ImageDimension;
+use TYPO3\CMS\Core\Resource\Processing\TaskInterface;
+
+class ImgixImageService extends ImageServiceAbstract
+{
+	public static function getIdentifier(): string
+	{
+		return 'imgix';
+	}
+
+	public function __construct(
+		protected readonly string $endpoint,
+		protected readonly UriSourceInterface $source,
+		protected readonly string $key,
+	) {
+	}
+
+	public function getEndpoint(): string
+	{
+		return $this->endpoint;
+	}
+
+	public function getKey(): ?string
+	{
+		return $this->key;
+	}
+
+	public function hasConfiguration(): bool
+	{
+		return filter_var($this->getEndpoint(), FILTER_VALIDATE_URL) !== false;
+	}
+
+	public function canProcessTask(TaskInterface $task): bool
+	{
+		return
+			$task->getSourceFile()->exists() &&
+			$task->getSourceFile()->getStorage()?->isPublic() &&
+			in_array($task->getName(), ['Preview', 'CropScaleMask'], true) &&
+			in_array($task->getSourceFile()->getMimeType(), [
+				'image/jpeg',
+				'image/png',
+				'image/gif',
+				'image/webp',
+				'image/heic',
+				'image/bmp',
+				'image/eps',
+				'image/tiff',
+				'application/pdf',
+			]);
+	}
+
+	public function processTask(TaskInterface $task): ImageServiceResult
+	{
+		$file = $task->getSourceFile();
+		$configuration = $task->getTargetFile()->getProcessingConfiguration();
+		$dimension = ImageDimension::fromProcessingTask($task);
+
+		$uri = new ImgixUri(
+			$this->getEndpoint(),
+			$this->getKey(),
+		);
+
+		$uri->setSource($this->source->getSource($file));
+
+		$fit = (static function ($configuration) {
+			switch (true) {
+				default:
+					return 'scale';
+
+				case str_ends_with((string) ($configuration['width'] ?? ''), 'c'):
+				case str_ends_with((string) ($configuration['height'] ?? ''), 'c'):
+					return 'crop';
+
+				case str_ends_with((string) ($configuration['width'] ?? ''), 'm'):
+				case str_ends_with((string) ($configuration['height'] ?? ''), 'm'):
+				case isset($configuration['maxWidth']):
+				case isset($configuration['maxHeight']):
+					return 'clip';
+			}
+		})($configuration);
+
+		$uri->setFit($fit);
+
+		$uri->setWidth((int) ($configuration['width'] ?? $configuration['maxWidth']));
+		$uri->setHeight((int) ($configuration['height'] ?? $configuration['maxHeight']));
+
+		if (isset($configuration['crop'])) {
+			$uri->setRect(
+				(int) ($configuration['crop']->getOffsetLeft()),
+				(int) ($configuration['crop']->getOffsetTop()),
+				(int) ($configuration['crop']->getWidth()),
+				(int) ($configuration['crop']->getHeight()),
+			);
+		}
+
+		return new ImageServiceResult(
+			$uri,
+			$dimension,
+		);
+	}
+}